Let Queue object to fetch messages, not the QueueManager
