+++ src/Queue/QueueManager.php
@@ -5,6 +5,5 @@ namespace Landingi\AwsBundle\Queue;
 
 interface QueueManager
 {
-    public function getMessages(): array;
     public function getQueues(): array;
 }
